#226 Add auth guard for Account component + fix duplicate sidebar
